python/kafka-python/requirements.txt
kafka-python-ng==2.2.3
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation-kafka-python==0.51b0
